+++ Tests/MassTransit.AzureServiceBus/MassTransit.AzureServiceBus.Application/DependencyInjection.cs
@@ -2,6 +2,9 @@ using System.Reflection;
 using FluentValidation;
 using Intent.RoslynWeaver.Attributes;
 using MassTransit.AzureServiceBus.Application.Common.Behaviours;
+using MassTransit.AzureServiceBus.Application.Common.Eventing;
+using MassTransit.AzureServiceBus.Application.IntegrationEventHandlers;
+using MassTransit.AzureServiceBus.Eventing.Messages;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -22,6 +25,7 @@ namespace MassTransit.AzureServiceBus.Application
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(EventBusPublishBehaviour<,>));
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehaviour<,>));
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(UnitOfWorkBehaviour<,>));
+            services.AddTransient<IIntegrationEventHandler<TestMessageEvent>, TestMessageEventHandler>();
             return services;
         }
     }