+++ Tests/MassTransit.AzureServiceBus/MassTransit.AzureServiceBus.Application/IntegrationEventHandlers/TestMessageEventHandler.cs
@@ -1,0 +1,31 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using MassTransit.AzureServiceBus.Application.Common.Eventing;
+using MassTransit.AzureServiceBus.Eventing.Messages;
+using Microsoft.Extensions.Logging;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.MassTransit.IntegrationEventHandlerImplementation", Version = "1.0")]
+
+namespace MassTransit.AzureServiceBus.Application.IntegrationEventHandlers
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class TestMessageEventHandler : IIntegrationEventHandler<TestMessageEvent>
+    {
+        private readonly ILogger<TestMessageEventHandler> _logger;
+
+        [IntentManaged(Mode.Ignore)]
+        public TestMessageEventHandler(ILogger<TestMessageEventHandler> logger)
+        {
+            _logger = logger;
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
+        public async Task HandleAsync(TestMessageEvent message, CancellationToken cancellationToken = default)
+        {
+            _logger.LogInformation("Message received: {Message}", message);
+        }
+    }
+}