+++ Tests/MassTransit.AzureServiceBus/MassTransit.AzureServiceBus.Application/IntegrationEvents/TestMessageEvent.cs
@@ -1,0 +1,12 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.Contracts.IntegrationEventMessage", Version = "1.0")]
+
+namespace MassTransit.AzureServiceBus.Eventing.Messages
+{
+    public record TestMessageEvent
+    {
+        public string Message { get; init; }
+    }
+}