+++ Tests/MassTransit.AzureServiceBus/MassTransit.AzureServiceBus.Application/Test/SendTest/SendTestCommandValidator.cs
@@ -1,0 +1,25 @@
+using System;
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "1.0")]
+
+namespace MassTransit.AzureServiceBus.Application.Test.SendTest
+{
+    public class SendTestCommandValidator : AbstractValidator<SendTestCommand>
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore, Signature = Mode.Merge)]
+        public SendTestCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [IntentManaged(Mode.Fully)]
+        private void ConfigureValidationRules()
+        {
+            RuleFor(v => v.Message)
+                .NotNull();
+        }
+    }
+}