+++ Tests/MassTransit.RabbitMQ/MassTransit.RabbitMQ.Application/Common/Behaviours/EventBusPublishBehaviour.cs
@@ -1,0 +1,30 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using MassTransit.RabbitMQ.Application.Common.Eventing;
+using MediatR;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.Behaviours.EventBusPublishBehaviour", Version = "1.0")]
+
+namespace MassTransit.RabbitMQ.Application.Common.Behaviours;
+
+public class EventBusPublishBehaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+where TRequest : IRequest<TResponse>
+{
+    private readonly IEventBus _eventBus;
+
+    public EventBusPublishBehaviour(IEventBus eventBus)
+    {
+        _eventBus = eventBus;
+    }
+
+    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+    {
+        var response = await next();
+
+        await _eventBus.FlushAllAsync(cancellationToken);
+
+        return response;
+    }
+}