+++ Tests/MassTransit.RabbitMQ/MassTransit.RabbitMQ.Application/Common/Behaviours/PerformanceBehaviour.cs
@@ -1,0 +1,54 @@
+using System.Diagnostics;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using MassTransit.RabbitMQ.Application.Common.Interfaces;
+using MediatR;
+using Microsoft.Extensions.Logging;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.Behaviours.PerformanceBehaviour", Version = "1.0")]
+
+namespace MassTransit.RabbitMQ.Application.Common.Behaviours
+{
+    public class PerformanceBehaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+        where TRequest : IRequest<TResponse>
+    {
+        private readonly Stopwatch _timer;
+        private readonly ILogger<TRequest> _logger;
+        private readonly ICurrentUserService _currentUserService;
+
+        public PerformanceBehaviour(
+            ILogger<TRequest> logger,
+            ICurrentUserService currentUserService)
+        {
+            _timer = new Stopwatch();
+
+            _logger = logger;
+            _currentUserService = currentUserService;
+        }
+
+        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+        {
+            _timer.Start();
+
+            var response = await next();
+
+            _timer.Stop();
+
+            var elapsedMilliseconds = _timer.ElapsedMilliseconds;
+
+            if (elapsedMilliseconds > 500)
+            {
+                var requestName = typeof(TRequest).Name;
+                var userId = _currentUserService.UserId ?? string.Empty;
+                var userName = _currentUserService.UserName ?? string.Empty;
+
+                _logger.LogWarning("MassTransit.RabbitMQ Long Running Request: {Name} ({ElapsedMilliseconds} milliseconds) {@UserId} {@UserName} {@Request}",
+                    requestName, elapsedMilliseconds, userId, userName, request);
+            }
+
+            return response;
+        }
+    }
+}