+++ Tests/MassTransit.RabbitMQ/MassTransit.RabbitMQ.Infrastructure/Configuration/MassTransitConfiguration.cs
@@ -1,0 +1,43 @@
+using System;
+using System.Reflection;
+using Intent.RoslynWeaver.Attributes;
+using MassTransit;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.MassTransit.MassTransitConfiguration", Version = "1.0")]
+
+namespace MassTransit.RabbitMQ.Infrastructure.Configuration
+{
+    public static class MassTransitConfiguration
+    {
+        public static void AddMassTransitConfiguration(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddMassTransit(x =>
+            {
+                x.SetKebabCaseEndpointNameFormatter();
+
+                AddConsumers(x);
+
+                x.UsingRabbitMq((context, cfg) =>
+                {
+                    cfg.UseMessageRetry(r => r.Interval(10, TimeSpan.FromSeconds(30)));
+
+                    cfg.Host(configuration["RabbitMq:Host"], configuration["RabbitMq:VirtualHost"], h =>
+                    {
+                        h.Username(configuration["RabbitMq:Username"]);
+                        h.Password(configuration["RabbitMq:Password"]);
+                    });
+
+                    cfg.ConfigureEndpoints(context);
+                });
+            });
+        }
+
+        private static void AddConsumers(IRegistrationConfigurator cfg)
+        {
+
+        }
+    }
+}